fix: добавляет форматирование даты отзыва, выделяет mode константы в отдельный файл
