+++ src/components/places-list/place-card/place-card.tsx
@@ -4,7 +4,8 @@ import classNames from 'classnames';
 import { PlaceCardMode } from '../../../types/common';
 import { Offer } from '../../../types/offers';
 
-import { PlaceCardModeOption, AppRoute } from '../../../const/const';
+import { AppRoute } from '../../../const/const';
+import { PlaceCardModeOption } from '../../../const/mode';
 import Rating from '../../shared/rating/rating';
 import BookmarkButton from '../../shared/bookmark-button/bookmark-button';
 import OfferPrice from '../../shared/offer-price/offer-price';
