+++ src/components/places-list/places-list.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Offers } from '../../types/offers';
 import { PlaceCardMode } from '../../types/common';
 
-import { PlaceCardModeOption } from '../../const/const';
+import { PlaceCardModeOption } from '../../const/mode';
 import { getFavoriteOffers } from '../../utils/filter-utils';
 import PlaceCard from './place-card/place-card';
