+++ src/components/shared/bookmark-button/bookmark-button.tsx
@@ -1,7 +1,7 @@
 import { BookmarkButtonMode } from '../../../types/common';
 import { Offer } from '../../../types/offers';
 
-import { BookmarkButtonModeOption } from '../../../const/const';
+import { BookmarkButtonModeOption } from '../../../const/mode';
 
 type BookmarkButtonProps = {
   offer: Offer;
